Performance improvement: Don't remove old attribute before setting new one if possible. Saw a noticeable performance hit for setting style attribute.

+++ src/lib/component-types/host/set-attributes.ts
@@ -45,7 +45,8 @@ function updateAttrInner(
       setAttribute(element, namespace, key, n)
     } //
     else if (o !== n) {
-      deleteAttribute(element, key, o)
+      if (key.startsWith('on')) deleteAttribute(element, key, o)
+
       setAttribute(element, namespace, key, n)
     }
   }
